move list to lagacy
