+++ legacy/src/main/java/net/kwmt27/jetpackcomposeplayground/legacy/feature/list/instagram/InstagramSearchListLayout.kt
@@ -1,4 +1,4 @@
-package net.kwmt27.jetpackcomposeplayground.list.instagram
+package net.kwmt27.jetpackcomposeplayground.legacy.feature.list.instagram
 
 import android.util.Log
 import androidx.compose.foundation.background
@@ -15,6 +15,7 @@ import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
@@ -30,8 +31,51 @@ import net.kwmt27.jetpackcomposeplayground.legacy.ui.theme.JetpackComposePlayGro
 
 private const val TAG = "InstagramSearchList"
 
+enum class MoviePosition {
+    LEFT, RIGHT
+}
+
+data class GridRowData(
+    val list: List<ItemData>,
+)
+
+sealed interface ItemData {
+    val id: Int
+
+    @Stable
+    data class ItemImage(
+        override val id: Int,
+        val list: List<String> = emptyList()
+    ) : ItemData
+
+    @Stable
+    data class ItemMovie(
+        override val id: Int,
+    ) : ItemData
+}
+
+data class GridListData(
+    val list: List<GridRowData>,
+) {
+    companion object {
+        fun createData(): GridListData {
+            return (1..1000).mapIndexed { index, i ->
+                if (index % 5 == 0 && index != 0) {
+                    ItemData.ItemMovie(i)
+                } else {
+                    ItemData.ItemImage(i)
+                }
+            }.chunked(5).map { items ->
+                GridRowData(items)
+            }.run {
+                GridListData(this)
+            }
+        }
+    }
+}
+
 @Composable
-fun SampleInstagramSearchListLayout3() {
+fun SampleInstagramSearchListLayout() {
     InstagramSearchListLayout(GridListData.createData())
 }
 
@@ -40,16 +84,17 @@ private fun InstagramSearchListLayout(
     gridListData: GridListData,
 ) {
     val listState = rememberLazyListState()
-    // OK例
-    val playMovieIndex by remember {
-        derivedStateOf {
-            if (listState.firstVisibleItemScrollOffset == 0) {
-                listState.firstVisibleItemIndex
-            } else {
-                listState.firstVisibleItemIndex + 1
-            }
+    // NG例
+    // listState.firstVisibleItemScrollOffsetやlistState.firstVisibleItemIndexはスクロールの状態によって変化するもので、
+    // この場合だと、firstVisibleItemScrollOffsetが0かそれ以外かでしか関心がないが、
+    // このコードだとスクロールのたびにRecomposeされてします。
+    val playMovieIndex =
+        if (listState.firstVisibleItemScrollOffset == 0) {
+            listState.firstVisibleItemIndex
+        } else {
+            listState.firstVisibleItemIndex + 1
         }
-    }
+
     BoxWithConstraints {
         LazyColumn(
             modifier = Modifier
@@ -67,7 +112,12 @@ private fun InstagramSearchListLayout(
                     gridRowData = gridRowData,
                     width = maxWidth / 3,
                     moviePosition = moviePosition,
-                    isPlay = { index == playMovieIndex }
+                    listState = listState,
+                    index = index,
+//                    // [isPlay]はスクロールの状態によって変化するため、
+//                    // Booleanを直接渡すと、recomposeされてしまう。
+//                    // そのため、関数を渡すことによりCompositionを防ぐ
+                    isPlay = index == playMovieIndex
                 )
             }
         }
@@ -79,8 +129,21 @@ private fun GridRowItem(
     gridRowData: GridRowData,
     width: Dp,
     moviePosition: MoviePosition,
-    isPlay: () -> Boolean,
+    listState: LazyListState,
+    index: Int,
+    isPlay: Boolean,
 ) {
+//    val playMovieIndex by remember {
+//        derivedStateOf {
+//            if (listState.firstVisibleItemScrollOffset == 0) {
+//                listState.firstVisibleItemIndex
+//            } else {
+//                listState.firstVisibleItemIndex + 1
+//            }
+//        }
+//    }
+//
+//    val isPlay = { playMovieIndex ==index }
     /**
      * [item][item][ItemImage]
      * [item][item] ---
@@ -140,9 +203,9 @@ private fun ItemImage(item: ItemData, width: Dp) {
 }
 
 @Composable
-private fun ItemMovie(item: ItemData, width: Dp, isPlay: () -> Boolean,) {
-    Log.d(TAG, "ItemMovie: id=${item.id},  isPlay=${isPlay()}")
-    val color = if (isPlay()) Color.Green else Color.Red
+private fun ItemMovie(item: ItemData, width: Dp, isPlay: Boolean) {
+    Log.d(TAG, "ItemMovie: id=${item.id},  isPlay=$isPlay")
+    val color = if (isPlay) Color.Green else Color.Red
     Column(
         modifier = Modifier
             .size(width = width, height = width * 2 + 1.dp)
@@ -154,7 +217,7 @@ private fun ItemMovie(item: ItemData, width: Dp, isPlay: () -> Boolean,) {
             text = "${item.id} 動画",
             style = TextStyle(fontSize = 20.sp, color = Color.White)
         )
-        val text = if (isPlay()) "Playing" else "Stop"
+        val text = if (isPlay) "Playing" else "Stop"
         Text(
             text = text,
             style = TextStyle(fontSize = 20.sp, color = Color.White)
@@ -198,8 +261,31 @@ private fun ItemMovie(item: ItemData, width: Dp, listState: LazyListState, index
 
 @Preview(showBackground = true, device = "spec:width=320dp,height=640dp")
 @Composable
-private fun PreviewInstagramSearchList3() {
+private fun PreviewInstagramSearchList() {
     JetpackComposePlayGroundTheme {
-        SampleInstagramSearchListLayout3()
+        InstagramSearchListLayout(GridListData.createData())
+    }
+}
+
+@Preview(showBackground = true, device = "spec:width=320dp,height=640dp")
+@Composable
+private fun PreviewGridItemImages() {
+    JetpackComposePlayGroundTheme {
+        BoxWithConstraints {
+            GridItemImages(
+                GridListData.createData().list.get(2).list.take(4).chunked(2),
+                320.dp / 3
+            )
+        }
+    }
+}
+
+@Preview(showBackground = true, device = "spec:width=320dp,height=640dp")
+@Composable
+private fun PreviewItemMovie() {
+    JetpackComposePlayGroundTheme {
+        BoxWithConstraints {
+            ItemMovie(GridListData.createData().list.get(2).list.last(), 320.dp / 3, isPlay = false)
+        }
     }
 }
